+++ adapters/AppLovin/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## AppLovin iOS Mediation Adapter Changelog
 
 #### Next version
-- Added new adapter API implementation for waterfall interstitial ad format.
+
+- Added new adapter API implementation for waterfall banner and interstitial ad
+  formats.
 
 #### [Version 13.6.2.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-13.6.2.0.zip)
 - Verified compatibility with AppLovin SDK 13.6.2.